Refactor reminder flow and remove prototype code
