+++ app/src/main/java/com/github/jimmy90109/geoalarm/data/PlaceReminderLimits.kt
@@ -1,3 +1,3 @@
 package com.github.jimmy90109.geoalarm.data
 
-const val MaxPlaceReminderAttachments = 5
+const val MaxPlaceReminderAttachments = 6
